ci(update-tools): run create-pull-request with delete-branch true
Attempting to work around a case where apprently the machine user's
branch receives updates to workflow files from upstream (i.e. they are
not part of the PR it is about to make, just updates its local branch),
but is not allowed to push them.

```
Pushing pull request branch to 'fork/create-pull-request/patch'
  /usr/bin/git push --force-with-lease fork create-pull-request/patch:refs/heads/create-pull-request/patch
  To https://github.com/bot-scop/wrun
   ! [remote rejected] create-pull-request/patch -> create-pull-request/patch (refusing to allow a Personal Access Token to create or update workflow `.github/workflows/check-generators.yaml` without `workflow` scope)
  error: failed to push some refs to 'https://github.com/bot-scop/wrun'
```